Refactor code to improve search functionality
- Updated the export file to include the new search feature
- Created a new debouncer and throttler class to improve search functionality
- Added a new service locator file to register the new search cubit and repo
- Created a new search screen with a search field and a list of search results
- Updated the API service to include a new search endpoint
- Updated the product details screen to use the new search screen when searching for products
- Updated the constants file to include a new search screen route

This commit refactors the code to improve the search functionality by adding a new search screen, debouncer, throttler, and service locator file. It also updates the API service to include a new search endpoint and the product details screen to use the new search screen when searching for products. Finally, it updates the constants file to include a new search screen route.